backend/requirements.txt
fastapi==0.123.0
uvicorn[standard]==0.38.0
google-generativeai>=0.8.3
python-dotenv==1.0.0
pydantic==2.10.5
